ft. mejora masiva
wazaaa

+++ src/main/java/com/grupoperuana/sistema/repositories/PermisoRepository.java
@@ -5,8 +5,9 @@ import com.grupoperuana.sistema.beans.Permiso;
 import java.util.Optional;
 import java.util.List;
 
-public interface PermisoRepository extends JpaRepository<Permiso, Long> {
+public interface PermisoRepository extends JpaRepository<Permiso, Integer> {
 
     Optional<Permiso> findByNombre(String nombre);
+
     List<Permiso> findByNombreIn(List<String> nombres);
 }
